LayoutTests/fast/scrolling/ios/smooth-scroll-animates-expected.txt

PASS window.scrollY < 500 is true
PASS scrollEventCount > 1 is true
PASS successfullyParsed is true

TEST COMPLETE
